05 May 2023: SleepActivity.java - Added functionality to only update RecyclerViewer if sleepConfidence has changed instead of an updated every second. Function checks if last element in array is different to the second to last element. The last element is sent from SleepTrackerService.java via the countdown boradcast.
SleepTrackerService.java - Fixed 'double pause' problem, if the countdown had started and paused due to user waking up, and the next 2 sleep confidence levels didn't resume the timer, the TextView in SleepActivity for the countdown timer reset to the original set value. An extra condition added for these circumstances which keeps the saved countdown value.